+++ src/libft.h
@@ -91,8 +91,8 @@ int		ft_sign(int n);
 char	*ft_itoa_base(int n, char *symbols);
 void	ft_putnbr_base_fd(int n, char *symbols, int fd);
 
-char	*ft_itoa_base_unsigned(unsigned int n, char *symbols);
-void	ft_putnbr_base_unsigned_fd(unsigned int n, char *symbols, int fd);
+char	*ft_itoa_base_unsigned(size_t n, char *symbols);
+void	ft_putnbr_base_unsigned_fd(size_t n, char *symbols, int fd);
 
 
 #endif
